+++ opening-hours-syntax/src/rules/mod.rs
@@ -5,7 +5,7 @@ use std::fmt::Display;
 use std::sync::Arc;
 
 use crate::normalize::frame::Bounded;
-use crate::normalize::paving::{Paving, Paving5D};
+use crate::normalize::paving::{Paving, Paving5D, UnpackFromBack};
 use crate::normalize::{canonical_to_seq, ruleseq_to_selector};
 use crate::sorted_vec::UniqueSortedVec;
 
@@ -72,8 +72,8 @@ impl OpeningHoursExpression {
             // If the rule is not explicitly targeting a closed kind, then it overrides
             // previous rules for the whole day.
             if rule.operator == RuleOperator::Normal && rule.kind != RuleKind::Closed {
-                let (_, day_selector) = selector.clone().into_unpack_front();
-                let full_day_selector = day_selector.dim_front([Bounded::bounds()]);
+                let mut full_day_selector = selector.clone();
+                full_day_selector.substitute_back([Bounded::bounds()]);
                 paving.set(&full_day_selector, &Default::default());
             }